split data into its own file
