+++ src/mocks/data/user-group/user-group.data.ts
@@ -1,46 +1,9 @@
-import { UmbEntityData } from '../entity.data.js';
+import { UserGroupResponseModel } from '@umbraco-cms/backoffice/backend-api';
 import {
 	UMB_USER_PERMISSION_DOCUMENT_CREATE,
 	UMB_USER_PERMISSION_DOCUMENT_DELETE,
 	UMB_USER_PERMISSION_DOCUMENT_READ,
 } from '@umbraco-cms/backoffice/document';
-import { UserGroupItemResponseModel, UserGroupResponseModel } from '@umbraco-cms/backoffice/backend-api';
-
-const createUserGroupItem = (item: UserGroupResponseModel): UserGroupItemResponseModel => {
-	return {
-		name: item.name,
-		id: item.id,
-		icon: item.icon,
-	};
-};
-
-// Temp mocked database
-class UmbUserGroupData extends UmbEntityData<UserGroupResponseModel> {
-	constructor(data: Array<UserGroupResponseModel>) {
-		super(data);
-	}
-
-	getItems(ids: Array<string>): Array<UserGroupItemResponseModel> {
-		const items = this.data.filter((item) => ids.includes(item.id ?? ''));
-		return items.map((item) => createUserGroupItem(item));
-	}
-
-	/**
-	 * Returns a list of permissions for the given user group ids
-	 * @param {string[]} userGroupIds
-	 * @return {*}  {string[]}
-	 * @memberof UmbUserGroupData
-	 */
-	getPermissions(userGroupIds: string[]): string[] {
-		const permissions = this.data
-			.filter((userGroup) => userGroupIds.includes(userGroup.id || ''))
-			.map((userGroup) => (userGroup.permissions?.length ? userGroup.permissions : []))
-			.flat();
-
-		// Remove duplicates
-		return [...new Set(permissions)];
-	}
-}
 
 export const data: Array<UserGroupResponseModel> = [
 	{
@@ -98,5 +61,3 @@ export const data: Array<UserGroupResponseModel> = [
 		permissions: [UMB_USER_PERMISSION_DOCUMENT_READ],
 	},
 ];
-
-export const umbUserGroupData = new UmbUserGroupData(data);
